requirements.txt
aws-cdk-lib
constructs
PyYAML